Fixed a warning.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@10664 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/TaskManager.cpp
@@ -132,7 +132,7 @@ bool TaskManager::areAllThreadsBusy()
 {
 	Lock lock(num_unfinished_tasks_mutex);
 
-	return num_unfinished_tasks >= threads.size();
+	return num_unfinished_tasks >= (int)threads.size();
 }
